infra/indexer/requirements.txt
kafka-python==2.0.2
opensearch-py==2.4.2
python-dotenv==1.0.1
